- Renamed class - updated for Room library
